add supported chain validation to form
